requirements.txt
cassandra-driver>=3.25.0
pytest
pytest-mock
python-dotenv
langchain
langchain-openai